auditworkshop/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
pgvector==0.3.2
sentence-transformers==3.1.1
pdfplumber==0.11.4
pymupdf==1.24.11
httpx==0.27.2
python-multipart==0.0.12
openpyxl==3.1.5
pandas==2.2.3
requests==2.32.3
psutil==6.0.0
sqlalchemy==2.0.35
alembic==1.13.1
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.1.0
